[BF16] BF16 simulator for AVX512 was done (#3424)

+++ inference-engine/tests/functional/plugin/cpu/bfloat16/conv_conv.cpp
@@ -100,7 +100,7 @@ protected:
         // performance counters
         expectedPrecisions["ADD_1"] = "FP32";
         expectedPrecisions["CONV_1"] = "BF16";
-        expectedPrecisions["CONV_2"] = "BF16";
+        expectedPrecisions["CONV_2"] = "FP32";
     }
 };
